some changes after merge
