+++ src/main/java/faang/school/achievement/handler/AbstractAchievementHandler.java
@@ -2,38 +2,37 @@ package faang.school.achievement.handler;
 
 import faang.school.achievement.model.Achievement;
 import faang.school.achievement.model.AchievementProgress;
-import faang.school.achievement.event.AlbumCreatedEvent;
-import faang.school.achievement.model.UserAchievement;
 import faang.school.achievement.service.AchievementCache;
 import faang.school.achievement.service.AchievementService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Async;
-import org.springframework.stereotype.Service;
 
 @Slf4j
-@Service
 @RequiredArgsConstructor
-public abstract class AbstractAchievementHandler implements EventHandler<AlbumCreatedEvent> {
+public abstract class AbstractAchievementHandler<T> implements EventHandler<T> {
 
-    protected final AchievementCache achievementCache;
-    protected final AchievementService achievementService;
+    private final AchievementService achievementService;
+    private final AchievementCache achievementCache;
 
-    @Async("eventHandlingTaskExecutor")
-    public void handleAchievement(AlbumCreatedEvent event, String titleAchievement) {
-        Achievement achievement = achievementCache.get(titleAchievement);
-        if (!achievementService.hasAchievement(event.getUserId(), achievement.getId())) {
-            achievementService.createProgressIfNecessary(event.getUserId(), achievement.getId());
-            AchievementProgress achievementProgress = achievementService.getProgress(event.getUserId(), achievement.getId());
-            achievementProgress.increment();
-            achievementService.saveProgress(achievementProgress);
-            if (achievementProgress.getCurrentPoints() >= achievement.getPoints()) {
-                UserAchievement userAchievement = new UserAchievement();
-                userAchievement.setUserId(event.getUserId());
-                userAchievement.setAchievement(achievement);
-                achievementService.giveAchievement(userAchievement);
-            }
+    @Async("achievementHandlingExecutor")
+    public void handleAchievement(long userId, String achievementTitle) {
+        log.info("Handling achievement titled '{}'", achievementTitle);
+        Achievement achievement = achievementCache.get(achievementTitle);
+        if (achievementService.hasAchievement(userId, achievement.getId())) {
+            log.info("User with ID={} already has achievement titled '{}'", userId, achievementTitle);
+            return;
         }
+        handleProgress(userId, achievement);
     }
 
+    private void handleProgress(long userId, Achievement achievement) {
+        achievementService.createProgressIfNecessary(userId, achievement.getId());
+        AchievementProgress achievementProgress = achievementService.getProgress(userId, achievement.getId());
+        achievementProgress.increment();
+        if (achievementProgress.getCurrentPoints() == achievement.getPoints()) {
+            achievementService.giveAchievement(userId, achievement);
+        }
+        achievementService.updateProgress(achievementProgress);
+    }
 }
