--- src/main/java/faang/school/achievement/handler/LibrarianAchievementHandler.java
@@ -1,20 +1,0 @@
-package faang.school.achievement.handler;
-
-import faang.school.achievement.event.AlbumCreatedEvent;
-import faang.school.achievement.service.AchievementCache;
-import faang.school.achievement.service.AchievementService;
-import org.springframework.stereotype.Service;
-
-@Service
-public class LibrarianAchievementHandler extends AbstractAchievementHandler {
-
-    public LibrarianAchievementHandler(AchievementCache achievementCache,
-                                       AchievementService achievementService) {
-        super(achievementCache, achievementService);
-    }
-
-    @Override
-    public void handleEvent(AlbumCreatedEvent event) {
-        handleAchievement(event, "LIBRARIAN");
-    }
-}
