+++ src/test/java/faang/school/achievement/handler/AbstractAchievementHandlerTest.java
@@ -2,7 +2,6 @@ package faang.school.achievement.handler;
 
 import faang.school.achievement.model.Achievement;
 import faang.school.achievement.model.AchievementProgress;
-import faang.school.achievement.event.AlbumCreatedEvent;
 import faang.school.achievement.service.AchievementCache;
 import faang.school.achievement.service.AchievementService;
 import org.junit.jupiter.api.Test;
@@ -11,80 +10,100 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
+import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
-
 
 @ExtendWith(MockitoExtension.class)
 public class AbstractAchievementHandlerTest {
 
     @Mock
-    private AchievementCache achievementCache;
-
-    @Mock
     private AchievementService achievementService;
 
+    @Mock
+    private AchievementCache achievementCache;
+
     @InjectMocks
-    private LibrarianAchievementHandler librarianAchievementHandler;
+    private TestAchievementHandler testEventHandler;
 
     @Test
-    void handleAchievementHasAchievementTest() {
-        AlbumCreatedEvent event = new AlbumCreatedEvent();
-        event.setUserId(1L);
-        String titleAchievement = "Test";
+    void handleAlreadyGivenAchievement() {
+        long userId = 1L;
+        long achievementId = 2L;
+        String achievementTitle = "title";
         Achievement achievement = new Achievement();
-        achievement.setTitle(titleAchievement);
-        achievement.setId(2L);
-        when(achievementCache.get("Test")).thenReturn(achievement);
-        when(achievementService.hasAchievement(event.getUserId(),achievement.getId())).thenReturn(true);
+        achievement.setTitle(achievementTitle);
+        achievement.setId(achievementId);
 
-        librarianAchievementHandler.handleAchievement(event, titleAchievement);
+        when(achievementCache.get(achievementTitle)).thenReturn(achievement);
+        when(achievementService.hasAchievement(userId, achievementId)).thenReturn(true);
 
-        verify(achievementService, times(0))
-                .createProgressIfNecessary(event.getUserId(), achievement.getId());
+        assertDoesNotThrow(() -> testEventHandler.handleAchievement(userId, achievementTitle));
+
+        verify(achievementCache, times(1)).get(achievementTitle);
+        verify(achievementService, times(1)).hasAchievement(userId, achievement.getId());
+        verify(achievementService, never()).createProgressIfNecessary(userId, achievement.getId());
+        verify(achievementService, never()).getProgress(userId, achievement.getId());
+        verify(achievementService, never()).giveAchievement(userId, achievement);
+        verify(achievementService, never()).updateProgress(any());
     }
 
     @Test
-    void handleAchievementNotEnoughPointsTest() {
-        AlbumCreatedEvent event = new AlbumCreatedEvent();
-        event.setUserId(1L);
-        String titleAchievement = "Test";
+    void handleNotFutureReceivedAchievement() {
+        long userId = 1L;
+        long achievementId = 2L;
+        long achievementPoints = 10L;
+        long currentPoints = 8L;
+        String achievementTitle = "title";
         Achievement achievement = new Achievement();
-        achievement.setTitle(titleAchievement);
-        achievement.setId(2L);
-        achievement.setPoints(10);
-        AchievementProgress achievementProgress = new AchievementProgress();
-        achievementProgress.setCurrentPoints(5);
-        when(achievementCache.get("Test")).thenReturn(achievement);
-        when(achievementService.hasAchievement(event.getUserId(),achievement.getId())).thenReturn(false);
-        when(achievementService.getProgress(event.getUserId(), achievement.getId())).thenReturn(achievementProgress);
+        achievement.setTitle(achievementTitle);
+        achievement.setId(achievementId);
+        achievement.setPoints(achievementPoints);
+        AchievementProgress progress = new AchievementProgress();
+        progress.setCurrentPoints(currentPoints);
 
-        librarianAchievementHandler.handleAchievement(event, titleAchievement);
+        when(achievementCache.get(achievementTitle)).thenReturn(achievement);
+        when(achievementService.hasAchievement(userId, achievementId)).thenReturn(false);
+        when(achievementService.getProgress(userId, achievementId)).thenReturn(progress);
 
-        verify(achievementService, times(0))
-                .giveAchievement(any());
+        assertDoesNotThrow(() -> testEventHandler.handleAchievement(userId, achievementTitle));
+
+        verify(achievementCache, times(1)).get(achievementTitle);
+        verify(achievementService, times(1)).hasAchievement(userId, achievement.getId());
+        verify(achievementService, times(1)).createProgressIfNecessary(userId, achievement.getId());
+        verify(achievementService, times(1)).getProgress(userId, achievement.getId());
+        verify(achievementService, never()).giveAchievement(userId, achievement);
+        verify(achievementService, times(1)).updateProgress(any());
     }
 
     @Test
-    void handleAchievementTest() {
-        AlbumCreatedEvent event = new AlbumCreatedEvent();
-        event.setUserId(1L);
-        String titleAchievement = "Test";
+    void handleFutureReceivedAchievement() {
+        long userId = 1L;
+        long achievementId = 2L;
+        long achievementPoints = 10L;
+        long currentPoints = 9L;
+        String achievementTitle = "title";
         Achievement achievement = new Achievement();
-        achievement.setTitle(titleAchievement);
-        achievement.setId(2L);
-        achievement.setPoints(10);
-        AchievementProgress achievementProgress = new AchievementProgress();
-        achievementProgress.setCurrentPoints(9);
-        when(achievementCache.get("Test")).thenReturn(achievement);
-        when(achievementService.hasAchievement(event.getUserId(),achievement.getId())).thenReturn(false);
-        when(achievementService.getProgress(event.getUserId(), achievement.getId())).thenReturn(achievementProgress);
+        achievement.setTitle(achievementTitle);
+        achievement.setId(achievementId);
+        achievement.setPoints(achievementPoints);
+        AchievementProgress progress = new AchievementProgress();
+        progress.setCurrentPoints(currentPoints);
 
-        librarianAchievementHandler.handleAchievement(event, titleAchievement);
+        when(achievementCache.get(achievementTitle)).thenReturn(achievement);
+        when(achievementService.hasAchievement(userId, achievementId)).thenReturn(false);
+        when(achievementService.getProgress(userId, achievementId)).thenReturn(progress);
 
-        verify(achievementService, times(1))
-                .giveAchievement(any());
+        assertDoesNotThrow(() -> testEventHandler.handleAchievement(userId, achievementTitle));
+
+        verify(achievementCache, times(1)).get(achievementTitle);
+        verify(achievementService, times(1)).hasAchievement(userId, achievement.getId());
+        verify(achievementService, times(1)).createProgressIfNecessary(userId, achievement.getId());
+        verify(achievementService, times(1)).getProgress(userId, achievement.getId());
+        verify(achievementService, times(1)).giveAchievement(userId, achievement);
+        verify(achievementService, times(1)).updateProgress(any());
     }
 }
