+++ src/test/java/faang/school/achievement/handler/TestAchievementHandler.java
@@ -3,9 +3,9 @@ package faang.school.achievement.handler;
 import faang.school.achievement.service.AchievementCache;
 import faang.school.achievement.service.AchievementService;
 
-public class TestEventHandler extends AbstractEventHandler<TestEvent> {
+public class TestAchievementHandler extends AbstractAchievementHandler<TestEvent> {
 
-    public TestEventHandler(AchievementService achievementService, AchievementCache achievementCache) {
+    public TestAchievementHandler(AchievementService achievementService, AchievementCache achievementCache) {
         super(achievementService, achievementCache);
     }
